refactor: moved filtering logic to useEvents hook
