+++ src/features/events/components/EventsList/index.tsx
@@ -1,33 +1,13 @@
-import { isAfter, isBefore } from 'date-fns'
 import type { FC } from 'react'
-import { useMemo, useState } from 'react'
+import { useState } from 'react'
 
 import { EventItem } from './parts/EventItem'
 import { NavigationFilter } from './parts/NavigationFilter'
 import { NavigationView } from './parts/NavigationView'
 import { Nav, List, ListItem } from './styled'
-import { ViewType, FilterType } from './types'
+import { ViewType } from './types'
 
-import { useEvents } from '../../hooks/useEvents'
-import type { Event } from '../../types'
-
-const sorts = {
-  asc: (a: Event, b: Event) => (a.startsAt < b.startsAt ? -1 : 1),
-  desc: (a: Event, b: Event) => (a.startsAt > b.startsAt ? -1 : 1),
-}
-
-const filters = {
-  future: (event: Event) => isAfter(new Date(event.startsAt), new Date()),
-  past: (event: Event) => isBefore(new Date(event.startsAt), new Date()),
-}
-
-const { ALL, FUTURE, PAST } = FilterType
-
-const listBuilders = {
-  [ALL]: (events: Event[]) => [...events].sort(sorts.asc),
-  [FUTURE]: (events: Event[]) => events.filter(filters.future).sort(sorts.asc),
-  [PAST]: (events: Event[]) => events.filter(filters.past).sort(sorts.desc),
-}
+import { useEvents, FilterType } from '../../hooks/useEvents'
 
 /**
  * Renders a list of events, with filtering/sorting/view type options.
@@ -36,8 +16,7 @@ export const EventsList: FC = () => {
   const [view, setView] = useState<ViewType>(ViewType.GRID)
   const [filter, setFilter] = useState<FilterType>(FilterType.ALL)
 
-  const { data, isLoading } = useEvents()
-  const events = useMemo(() => listBuilders[filter](data), [data, filter])
+  const { events, isLoading } = useEvents(filter)
 
   return (
     <>
