+++ src/features/events/components/EventsList/types.ts
@@ -2,9 +2,3 @@ export enum ViewType {
   GRID = 'GRID',
   LIST = 'LIST',
 }
-
-export enum FilterType {
-  ALL = 'ALL',
-  FUTURE = 'FUTURE',
-  PAST = 'PAST',
-}
